fix: fpm fails randomly
FPM would randomly fail to build deb and rpm packages. It apparently
happens because fpm (or some tool it uses) shares state between runs
(possibly by overriding files that are going inside the archive).

The error `file changed as we read it` was happening, which is a `tar`
error.

I don't know how to fix this, but, in order to make goreleaser more
stable, I'll disable the concurrency here for now.

closes #333